+++ src/lib/Filesystem/Node/File.php
@@ -104,8 +104,6 @@ class File extends AbstractNode implements IFile
 
     /**
      * Read content and return ressource.
-     *
-     * @return resource
      */
     public function get()
     {
@@ -126,8 +124,6 @@ class File extends AbstractNode implements IFile
 
     /**
      * Copy node.
-     *
-     * @param string $recursion
      */
     public function copyTo(Collection $parent, int $conflict = NodeInterface::CONFLICT_NOACTION, ?string $recursion = null, bool $recursion_first = true): NodeInterface
     {
@@ -200,7 +196,6 @@ class File extends AbstractNode implements IFile
         }
 
         $current = $this->version;
-        $new = $this->increaseVersion();
 
         $v = array_search($version, array_column($this->history, 'version'), true);
         if (false === $v) {
@@ -208,9 +203,10 @@ class File extends AbstractNode implements IFile
         }
 
         $file = $this->history[$v]['storage'];
+        $latest = $this->version + 1;
 
         $this->history[] = [
-            'version' => $new,
+            'version' => $latest,
             'changed' => $this->changed,
             'user' => $this->owner,
             'type' => self::HISTORY_RESTORE,
@@ -223,6 +219,7 @@ class File extends AbstractNode implements IFile
         ];
 
         try {
+            $new = $this->increaseVersion();
             $this->deleted = false;
             $this->version = $new;
             $this->storage = $this->history[$v]['storage'];
@@ -267,8 +264,6 @@ class File extends AbstractNode implements IFile
      *
      * Actually the node will not be deleted (Just set a delete flag), set $force=true to
      * delete finally
-     *
-     * @param string $recursion
      */
     public function delete(bool $force = false, ?string $recursion = null, bool $recursion_first = true): bool
     {
@@ -428,8 +423,6 @@ class File extends AbstractNode implements IFile
 
     /**
      * Get hash.
-     *
-     * @return string
      */
     public function getHash(): ?string
     {
@@ -546,7 +539,7 @@ class File extends AbstractNode implements IFile
         $max = $this->_fs->getServer()->getMaxFileVersion();
         if (count($this->history) >= $max) {
             $del = key($this->history);
-            $this->_logger->debug('history limit ['.$max.'] reached, remove oldest version ['.$del.'] from file ['.$this->_id.']', [
+            $this->_logger->debug('history limit ['.$max.'] reached, remove oldest version ['.$this->history[$del]['version'].'] from file ['.$this->_id.']', [
                 'category' => get_class($this),
             ]);