+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 **Date**: Mon Oct 22 10:32:56 CEST 2018
 
 * CORE: [CHANGE] Upgraded micro-auth to v0.0.7 (Accepts access_token from query string), see gyselroth/balloon-client-web #87
+* CORE: [FIX] Balloon\Filesystem\Exception\NotFound storage blob is gone exception after max_version reached #209
 
 
 ## 2.1.3